Suggesting a minor change to remove the check that sets SWRadAtm = 0._dp when cosZenith <= 0._dp these checks could be done by the user during pre-processing rather than at run time.
Line 165 was changed from: if(cosZenith <= 0._dp .or. SWRadAtm < 0._dp) SWRadAtm = 0._dp
to:   if(SWRadAtm < 0._dp) SWRadAtm = 0._dp

Thus, any measured shortwave radiation at times when the sun is at or below the horizon will be considered valid (and diffuse).
Note, it will be up to the user to ensure that nighttime shortwave radiation values are realistic (i.e., not positive).