Add nc_file_close calls to address potential memory leaks, and addressed conditional on unintialized variable.
I added close calls to the read_param.f90, and read_icond.f90 files. This does not really affect Summa as it does not enter these routines again. This does affect Summa-Actors as it computes batches and will enter the routine upon the next batch.

I also fixed an issue that can potentially cause undefined behaviour. In varSubStep.f90 the nrgFluxModified variable could end up uninitaialized and used as part of a conditional if checkMassBalance was false. I just initalized it to .false., but this may need more review to ensure this is correct.